function and import refactoring
