+++ visual_behavior/data_access/utilities.py
@@ -1,5 +1,14 @@
 import os
+import pandas as pd
+import json
+import shutil
 from visual_behavior.data_access import loading
+from visual_behavior.ophys.io.lims_database import LimsDatabase
+from visual_behavior.ophys.sync.sync_dataset import Dataset as SyncDataset
+from visual_behavior.ophys.sync.process_sync import filter_digital, calculate_delay
+
+import logging
+logger = logging.getLogger(__name__)
 
 
 # CONVENIENCE FUNCTIONS TO GET VARIOUS INFORMATION #
@@ -220,3 +229,163 @@ def get_timestamps(lims_data, analysis_dir):
     sync_data = get_sync_data(lims_data, analysis_dir, use_acq_trigger)
     timestamps = pd.DataFrame(sync_data)
     return timestamps
+
+
+def get_sync_path(lims_data, analysis_dir):
+    #    import shutil
+    ophys_session_dir = get_ophys_session_dir(lims_data)
+    # analysis_dir = get_analysis_dir(lims_data)
+
+    # First attempt
+    sync_file = [file for file in os.listdir(ophys_session_dir) if 'sync' in file]
+    if len(sync_file) > 0:
+        sync_file = sync_file[0]
+    else:
+        json_path = [file for file in os.listdir(ophys_session_dir) if '_platform.json' in file][0]
+        with open(os.path.join(ophys_session_dir, json_path)) as pointer_json:
+            json_data = json.load(pointer_json)
+            sync_file = json_data['sync_file']
+    sync_path = os.path.join(ophys_session_dir, sync_file)
+
+    if sync_file not in os.listdir(analysis_dir):
+        logger.info('moving %s to analysis dir', sync_file)  # flake8: noqa: E999
+        #        print(sync_path, os.path.join(analysis_dir, sync_file))
+        try:
+            shutil.copy2(sync_path, os.path.join(analysis_dir, sync_file))
+        except:  # NOQA E722
+            print('shutil.copy2 gave an error perhaps related to copying stat data... passing!')
+            pass
+    return sync_path
+
+
+def get_sync_data(lims_data, analysis_dir, use_acq_trigger):
+    logger.info('getting sync data')
+    sync_path = get_sync_path(lims_data, analysis_dir)
+    sync_dataset = SyncDataset(sync_path)
+    # Handle mesoscope missing labels
+    try:
+        sync_dataset.get_rising_edges('2p_vsync')
+    except ValueError:
+        sync_dataset.line_labels = ['2p_vsync', '', 'stim_vsync', '', 'photodiode', 'acq_trigger', '', '',
+                                    'behavior_monitoring', 'eye_tracking', '', '', '', '', '', '', '', '', '', '', '',
+                                    '', '', '', '', '', '', '', '', '', '', 'lick_sensor']
+        sync_dataset.meta_data['line_labels'] = sync_dataset.line_labels
+
+    meta_data = sync_dataset.meta_data
+    sample_freq = meta_data['ni_daq']['counter_output_freq']
+    # 2P vsyncs
+    vs2p_r = sync_dataset.get_rising_edges('2p_vsync')
+    vs2p_f = sync_dataset.get_falling_edges(
+        '2p_vsync', )  # new sync may be able to do units = 'sec', so conversion can be skipped
+    vs2p_rsec = vs2p_r / sample_freq
+    vs2p_fsec = vs2p_f / sample_freq
+    if use_acq_trigger:  # if 2P6, filter out solenoid artifacts
+        vs2p_r_filtered, vs2p_f_filtered = filter_digital(vs2p_rsec, vs2p_fsec, threshold=0.01)
+        frames_2p = vs2p_r_filtered
+    else:  # dont need to filter out artifacts in pipeline data
+        frames_2p = vs2p_rsec
+    # use rising edge for Scientifica, falling edge for Nikon http://confluence.corp.alleninstitute.org/display/IT/Ophys+Time+Sync
+    # Convert to seconds - skip if using units in get_falling_edges, otherwise convert before doing filter digital
+    # vs2p_rsec = vs2p_r / sample_freq
+    # frames_2p = vs2p_rsec
+    # stimulus vsyncs
+    # vs_r = d.get_rising_edges('stim_vsync')
+    vs_f = sync_dataset.get_falling_edges('stim_vsync')
+    # convert to seconds
+    # vs_r_sec = vs_r / sample_freq
+    vs_f_sec = vs_f / sample_freq
+    # vsyncs = vs_f_sec
+    # add display lag
+    monitor_delay = calculate_delay(sync_dataset, vs_f_sec, sample_freq)
+    vsyncs = vs_f_sec + monitor_delay  # this should be added, right!?
+    # line labels are different on 2P6 and production rigs - need options for both
+    if 'lick_times' in meta_data['line_labels']:
+        lick_times = sync_dataset.get_rising_edges('lick_1') / sample_freq
+    elif 'lick_sensor' in meta_data['line_labels']:
+        lick_times = sync_dataset.get_rising_edges('lick_sensor') / sample_freq
+    else:
+        lick_times = None
+    if '2p_trigger' in meta_data['line_labels']:
+        trigger = sync_dataset.get_rising_edges('2p_trigger') / sample_freq
+    elif 'acq_trigger' in meta_data['line_labels']:
+        trigger = sync_dataset.get_rising_edges('acq_trigger') / sample_freq
+    if 'stim_photodiode' in meta_data['line_labels']:
+        stim_photodiode = sync_dataset.get_rising_edges('stim_photodiode') / sample_freq
+    elif 'photodiode' in meta_data['line_labels']:
+        stim_photodiode = sync_dataset.get_rising_edges('photodiode') / sample_freq
+    if 'cam1_exposure' in meta_data['line_labels']:
+        eye_tracking = sync_dataset.get_rising_edges('cam1_exposure') / sample_freq
+    elif 'cam1' in meta_data['line_labels']:
+        eye_tracking = sync_dataset.get_rising_edges('cam1') / sample_freq
+    elif 'eye_tracking' in meta_data['line_labels']:
+        eye_tracking = sync_dataset.get_rising_edges('eye_tracking') / sample_freq
+    if 'cam2_exposure' in meta_data['line_labels']:
+        behavior_monitoring = sync_dataset.get_rising_edges('cam2_exposure') / sample_freq
+    elif 'cam2' in meta_data['line_labels']:
+        behavior_monitoring = sync_dataset.get_rising_edges('cam2') / sample_freq
+    elif 'behavior_monitoring' in meta_data['line_labels']:
+        behavior_monitoring = sync_dataset.get_rising_edges('behavior_monitoring') / sample_freq
+    # some experiments have 2P frames prior to stimulus start - restrict to timestamps after trigger for 2P6 only
+    if use_acq_trigger:
+        frames_2p = frames_2p[frames_2p > trigger[0]]
+    # print(len(frames_2p))
+    if lims_data.rig.values[0][0] == 'M':  # if Mesoscope
+        print('resampling mesoscope 2P frame times')
+        roi_group = get_roi_group(lims_data)  # get roi_group order
+        frames_2p = frames_2p[roi_group::4]  # resample sync times
+    # print(len(frames_2p))
+    logger.info('stimulus frames detected in sync: {}'.format(len(vsyncs)))
+    logger.info('ophys frames detected in sync: {}'.format(len(frames_2p)))
+    # put sync data in format to be compatible with downstream analysis
+    times_2p = {'timestamps': frames_2p}
+    times_vsync = {'timestamps': vsyncs}
+    times_lick = {'timestamps': lick_times}
+    times_trigger = {'timestamps': trigger}
+    times_eye_tracking = {'timestamps': eye_tracking}
+    times_behavior_monitoring = {'timestamps': behavior_monitoring}
+    times_stim_photodiode = {'timestamps': stim_photodiode}
+    sync_data = {'ophys_frames': times_2p,
+                 'stimulus_frames': times_vsync,
+                 'lick_times': times_lick,
+                 'eye_tracking': times_eye_tracking,
+                 'behavior_monitoring': times_behavior_monitoring,
+                 'stim_photodiode': times_stim_photodiode,
+                 'ophys_trigger': times_trigger,
+                 }
+    return sync_data
+
+
+def get_ophys_session_dir(lims_data):
+    ophys_session_dir = lims_data.ophys_session_dir.values[0]
+    return ophys_session_dir
+
+
+def get_ophys_experiment_dir(lims_data):
+    lims_id = get_lims_id(lims_data)
+    ophys_session_dir = get_ophys_session_dir(lims_data)
+    ophys_experiment_dir = os.path.join(ophys_session_dir, 'ophys_experiment_' + str(lims_id))
+    return ophys_experiment_dir
+
+
+def get_roi_group(lims_data):
+    experiment_id = int(lims_data.experiment_id.values[0])
+    ophys_session_dir = get_ophys_session_dir(lims_data)
+    import json
+    json_file = [file for file in os.listdir(ophys_session_dir) if ('SPLITTING' in file) and ('input.json' in file)]
+    json_path = os.path.join(ophys_session_dir, json_file[0])
+    with open(json_path, 'r') as w:
+        jin = json.load(w)
+    # figure out which roi_group the current experiment belongs to
+    # plane_data = pd.DataFrame()
+    for i, roi_group in enumerate(range(len(jin['plane_groups']))):
+        group = jin['plane_groups'][roi_group]['ophys_experiments']
+        for j, plane in enumerate(range(len(group))):
+            expt_id = int(group[plane]['experiment_id'])
+            if expt_id == experiment_id:
+                expt_roi_group = i
+    return expt_roi_group
+
+
+def get_lims_id(lims_data):
+    lims_id = lims_data.lims_id.values[0]
+    return lims_id
